+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 ## 3.0.1
+### Added
+- Possibility to specify hidden return value in `Terminal.interactiveSelectList`, `Terminal.interactiveMultiSelectList`, and `InteractiveSelectListBuilder`.
 ### Fixed
 - Fixed terminal size detection in the `mordant-jvm-ffm` module on macOS [(#238)](https://github.com/ajalt/mordant/issues/238)
 - Fixed proguard filename duplication [(#232)](https://github.com/ajalt/mordant/issues/232)